Remove hardcoded handling of the AccountModel. Read properties from System__Comobject correctly
Hardcoded AccountModel.dll and AccountImpl.dll meant that any other example will fail. Implemented functions that read the property values from a System__ComObject so that the correct dll is used for exploration.
Added code from ILSpy as this seems more reliable.
End-to-End test now runs! Open AccountModel, build it and explore SlicedModelProgram or TestSuite and we now match the behavior of the MS Built SE